+++ bindinfo/session_handle_test.go
@@ -32,58 +32,6 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestGlobalAndSessionBindingBothExist(t *testing.T) {
-	store, _, clean := testkit.CreateMockStoreAndDomain(t)
-	defer clean()
-
-	tk := testkit.NewTestKit(t, store)
-
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t1")
-	tk.MustExec("drop table if exists t2")
-	tk.MustExec("create table t1(id int)")
-	tk.MustExec("create table t2(id int)")
-	require.True(t, tk.HasPlan("SELECT * from t1,t2 where t1.id = t2.id", "HashJoin"))
-	require.True(t, tk.HasPlan("SELECT  /*+ TIDB_SMJ(t1, t2) */  * from t1,t2 where t1.id = t2.id", "MergeJoin"))
-
-	tk.MustExec("create global binding for SELECT * from t1,t2 where t1.id = t2.id using SELECT  /*+ TIDB_SMJ(t1, t2) */  * from t1,t2 where t1.id = t2.id")
-
-	// Test bindingUsage, which indicates how many times the binding is used.
-	metrics.BindUsageCounter.Reset()
-	require.True(t, tk.HasPlan("SELECT * from t1,t2 where t1.id = t2.id", "MergeJoin"))
-	pb := &dto.Metric{}
-	err := metrics.BindUsageCounter.WithLabelValues(metrics.ScopeGlobal).Write(pb)
-	require.Nil(t, err)
-	require.Equal(t, float64(1), pb.GetCounter().GetValue())
-
-	// Test 'tidb_use_plan_baselines'
-	tk.MustExec("set @@tidb_use_plan_baselines = 0")
-	require.True(t, tk.HasPlan("SELECT * from t1,t2 where t1.id = t2.id", "HashJoin"))
-	tk.MustExec("set @@tidb_use_plan_baselines = 1")
-
-	// Test 'drop global binding'
-	require.True(t, tk.HasPlan("SELECT * from t1,t2 where t1.id = t2.id", "MergeJoin"))
-	tk.MustExec("drop global binding for SELECT * from t1,t2 where t1.id = t2.id")
-	require.True(t, tk.HasPlan("SELECT * from t1,t2 where t1.id = t2.id", "HashJoin"))
-
-	// Test the case when global and session binding both exist
-	// PART1 : session binding should totally cover global binding
-	// use merge join as session binding here since the optimizer will choose hash join for this stmt in default
-	tk.MustExec("create global binding for SELECT * from t1,t2 where t1.id = t2.id using SELECT  /*+ TIDB_HJ(t1, t2) */  * from t1,t2 where t1.id = t2.id")
-	require.True(t, tk.HasPlan("SELECT * from t1,t2 where t1.id = t2.id", "HashJoin"))
-	tk.MustExec("create binding for SELECT * from t1,t2 where t1.id = t2.id using SELECT  /*+ TIDB_SMJ(t1, t2) */  * from t1,t2 where t1.id = t2.id")
-	require.True(t, tk.HasPlan("SELECT * from t1,t2 where t1.id = t2.id", "MergeJoin"))
-	tk.MustExec("drop global binding for SELECT * from t1,t2 where t1.id = t2.id")
-	require.True(t, tk.HasPlan("SELECT * from t1,t2 where t1.id = t2.id", "MergeJoin"))
-
-	// PART2 : the dropped session binding should continue to block the effect of global binding
-	tk.MustExec("create global binding for SELECT * from t1,t2 where t1.id = t2.id using SELECT  /*+ TIDB_SMJ(t1, t2) */  * from t1,t2 where t1.id = t2.id")
-	tk.MustExec("drop binding for SELECT * from t1,t2 where t1.id = t2.id")
-	require.True(t, tk.HasPlan("SELECT * from t1,t2 where t1.id = t2.id", "HashJoin"))
-	tk.MustExec("drop global binding for SELECT * from t1,t2 where t1.id = t2.id")
-	require.True(t, tk.HasPlan("SELECT * from t1,t2 where t1.id = t2.id", "HashJoin"))
-}
-
 func TestSessionBinding(t *testing.T) {
 	store, dom, clean := testkit.CreateMockStoreAndDomain(t)
 	defer clean()
@@ -107,15 +55,15 @@ func TestSessionBinding(t *testing.T) {
 
 		if testSQL.overlaySQL != "" {
 			_, err = tk.Exec("create session " + testSQL.overlaySQL)
-			require.Nil(t, err)
+			require.NoError(t, err)
 		}
 
 		pb := &dto.Metric{}
 		err = metrics.BindTotalGauge.WithLabelValues(metrics.ScopeSession, bindinfo.Using).Write(pb)
-		require.Nil(t, err)
+		require.NoError(t, err)
 		require.Equal(t, float64(1), pb.GetGauge().GetValue())
 		err = metrics.BindMemoryUsage.WithLabelValues(metrics.ScopeSession, bindinfo.Using).Write(pb)
-		require.Nil(t, err)
+		require.NoError(t, err)
 		require.Equal(t, testSQL.memoryUsage, pb.GetGauge().GetValue())
 
 		handle := tk.Session().Value(bindinfo.SessionBindInfoKeyType).(*bindinfo.SessionHandle)
@@ -132,17 +80,17 @@ func TestSessionBinding(t *testing.T) {
 		require.NotNil(t, bind.UpdateTime)
 
 		rs, err := tk.Exec("show global bindings")
-		require.Nil(t, err)
+		require.NoError(t, err)
 		chk := rs.NewChunk()
 		err = rs.Next(context.TODO(), chk)
-		require.Nil(t, err)
+		require.NoError(t, err)
 		require.Equal(t, 0, chk.NumRows())
 
 		rs, err = tk.Exec("show session bindings")
-		require.Nil(t, err)
+		require.NoError(t, err)
 		chk = rs.NewChunk()
 		err = rs.Next(context.TODO(), chk)
-		require.Nil(t, err)
+		require.NoError(t, err)
 		require.Equal(t, 1, chk.NumRows())
 		row := chk.GetRow(0)
 		require.Equal(t, testSQL.originSQL, row.GetString(0))
@@ -155,17 +103,17 @@ func TestSessionBinding(t *testing.T) {
 		require.NotNil(t, row.GetString(7))
 
 		_, err = tk.Exec("drop session " + testSQL.dropSQL)
-		require.Nil(t, err)
+		require.NoError(t, err)
 		bindData = handle.GetBindRecord(testSQL.originSQL, "test")
 		require.NotNil(t, bindData)
 		require.Equal(t, testSQL.originSQL, bindData.OriginalSQL)
-		require.Equal(t, 0, len(bindData.Bindings))
+		require.Len(t, bindData.Bindings, 0)
 
 		err = metrics.BindTotalGauge.WithLabelValues(metrics.ScopeSession, bindinfo.Using).Write(pb)
-		require.Nil(t, err)
+		require.NoError(t, err)
 		require.Equal(t, float64(0), pb.GetGauge().GetValue())
 		err = metrics.BindMemoryUsage.WithLabelValues(metrics.ScopeSession, bindinfo.Using).Write(pb)
-		require.Nil(t, err)
+		require.NoError(t, err)
 		require.Equal(t, float64(0), pb.GetGauge().GetValue())
 	}
 }
@@ -186,36 +134,36 @@ func TestBaselineDBLowerCase(t *testing.T) {
 	tk.MustExec("update t set a = a + 1")
 	tk.MustExec("admin capture bindings")
 	rows := tk.MustQuery("show global bindings").Rows()
-	require.Equal(t, 1, len(rows))
+	require.Len(t, rows, 1)
 	require.Equal(t, "update `spm` . `t` set `a` = `a` + ?", rows[0][0])
 	// default_db should have lower case.
 	require.Equal(t, "spm", rows[0][2])
 	tk.MustExec("drop global binding for update t set a = a + 1")
 	rows = tk.MustQuery("show global bindings").Rows()
 	// DROP GLOBAL BINGING should remove the binding even if we are in SPM database.
-	require.Equal(t, 0, len(rows))
+	require.Len(t, rows, 0)
 
 	tk.MustExec("create global binding for select * from t using select * from t")
 	rows = tk.MustQuery("show global bindings").Rows()
-	require.Equal(t, 1, len(rows))
+	require.Len(t, rows, 1)
 	require.Equal(t, "select * from `spm` . `t`", rows[0][0])
 	// default_db should have lower case.
 	require.Equal(t, "spm", rows[0][2])
 	tk.MustExec("drop global binding for select * from t")
 	rows = tk.MustQuery("show global bindings").Rows()
 	// DROP GLOBAL BINGING should remove the binding even if we are in SPM database.
-	require.Equal(t, 0, len(rows))
+	require.Len(t, rows, 0)
 
 	tk.MustExec("create session binding for select * from t using select * from t")
 	rows = tk.MustQuery("show session bindings").Rows()
-	require.Equal(t, 1, len(rows))
+	require.Len(t, rows, 1)
 	require.Equal(t, "select * from `spm` . `t`", rows[0][0])
 	// default_db should have lower case.
 	require.Equal(t, "spm", rows[0][2])
 	tk.MustExec("drop session binding for select * from t")
 	rows = tk.MustQuery("show session bindings").Rows()
 	// DROP SESSION BINGING should remove the binding even if we are in SPM database.
-	require.Equal(t, 0, len(rows))
+	require.Len(t, rows, 0)
 
 	utilCleanBindingEnv(tk, dom)
 
@@ -227,14 +175,14 @@ func TestBaselineDBLowerCase(t *testing.T) {
 	))
 	tk.MustExec("admin reload bindings")
 	rows = tk.MustQuery("show global bindings").Rows()
-	require.Equal(t, 1, len(rows))
+	require.Len(t, rows, 1)
 	require.Equal(t, "select * from `spm` . `t`", rows[0][0])
 	// default_db should have lower case.
 	require.Equal(t, "spm", rows[0][2])
 	tk.MustExec("drop global binding for select * from t")
 	rows = tk.MustQuery("show global bindings").Rows()
 	// DROP GLOBAL BINGING should remove the binding even if we are in SPM database.
-	require.Equal(t, 0, len(rows))
+	require.Len(t, rows, 0)
 
 	utilCleanBindingEnv(tk, dom)
 	// Simulate existing bindings with upper case default_db.
@@ -245,13 +193,13 @@ func TestBaselineDBLowerCase(t *testing.T) {
 	))
 	tk.MustExec("admin reload bindings")
 	rows = tk.MustQuery("show global bindings").Rows()
-	require.Equal(t, 1, len(rows))
+	require.Len(t, rows, 1)
 	require.Equal(t, "select * from `spm` . `t`", rows[0][0])
 	// default_db should have lower case.
 	require.Equal(t, "spm", rows[0][2])
 	tk.MustExec("create global binding for select * from t using select * from t")
 	rows = tk.MustQuery("show global bindings").Rows()
-	require.Equal(t, 1, len(rows))
+	require.Len(t, rows, 1)
 	require.Equal(t, "select * from `spm` . `t`", rows[0][0])
 	// default_db should have lower case.
 	require.Equal(t, "spm", rows[0][2])
@@ -262,7 +210,7 @@ func TestBaselineDBLowerCase(t *testing.T) {
 }
 
 func TestShowGlobalBindings(t *testing.T) {
-	store, _, clean := testkit.CreateMockStoreAndDomain(t)
+	store, clean := testkit.CreateMockStore(t)
 	defer clean()
 
 	tk := testkit.NewTestKit(t, store)
@@ -274,7 +222,7 @@ func TestShowGlobalBindings(t *testing.T) {
 	tk.MustExec("create table t0(a int, b int, key(a))")
 	require.True(t, tk.Session().Auth(&auth.UserIdentity{Username: "root", Hostname: "%"}, nil, nil))
 	rows := tk.MustQuery("show global bindings").Rows()
-	require.Equal(t, 0, len(rows))
+	require.Len(t, rows, 0)
 	// Simulate existing bindings in the mysql.bind_info.
 	tk.MustExec("insert into mysql.bind_info values('select * from `spm` . `t`', 'select * from `spm` . `t` USE INDEX (`a`)', 'SPM', 'using', '2000-01-01 09:00:00', '2000-01-01 09:00:00', '', '','" +
 		bindinfo.Manual + "')")
@@ -286,7 +234,7 @@ func TestShowGlobalBindings(t *testing.T) {
 		bindinfo.Manual + "')")
 	tk.MustExec("admin reload bindings")
 	rows = tk.MustQuery("show global bindings").Rows()
-	require.Equal(t, 4, len(rows))
+	require.Len(t, rows, 4)
 	require.Equal(t, "select * from `spm` . `t0`", rows[0][0])
 	require.Equal(t, "2000-01-04 09:00:00.000", rows[0][5])
 	require.Equal(t, "select * from `spm` . `t0`", rows[1][0])
@@ -297,13 +245,13 @@ func TestShowGlobalBindings(t *testing.T) {
 	require.Equal(t, "2000-01-01 09:00:00.000", rows[3][5])
 
 	rows = tk.MustQuery("show session bindings").Rows()
-	require.Equal(t, 0, len(rows))
+	require.Len(t, rows, 0)
 	tk.MustExec("create session binding for select a from t using select a from t")
 	tk.MustExec("create session binding for select a from t0 using select a from t0")
 	tk.MustExec("create session binding for select b from t using select b from t")
 	tk.MustExec("create session binding for select b from t0 using select b from t0")
 	rows = tk.MustQuery("show session bindings").Rows()
-	require.Equal(t, 4, len(rows))
+	require.Len(t, rows, 4)
 	require.Equal(t, "select `b` from `spm` . `t0`", rows[0][0])
 	require.Equal(t, "select `b` from `spm` . `t`", rows[1][0])
 	require.Equal(t, "select `a` from `spm` . `t0`", rows[2][0])
@@ -311,7 +259,7 @@ func TestShowGlobalBindings(t *testing.T) {
 }
 
 func TestDuplicateBindings(t *testing.T) {
-	store, _, clean := testkit.CreateMockStoreAndDomain(t)
+	store, clean := testkit.CreateMockStore(t)
 	defer clean()
 
 	tk := testkit.NewTestKit(t, store)
@@ -320,27 +268,27 @@ func TestDuplicateBindings(t *testing.T) {
 	tk.MustExec("create table t(a int, b int, index idx(a))")
 	tk.MustExec("create global binding for select * from t using select * from t use index(idx);")
 	rows := tk.MustQuery("show global bindings").Rows()
-	require.Equal(t, 1, len(rows))
+	require.Len(t, rows, 1)
 	createTime := rows[0][4]
 	time.Sleep(1000000)
 	tk.MustExec("create global binding for select * from t using select * from t use index(idx);")
 	rows = tk.MustQuery("show global bindings").Rows()
-	require.Equal(t, 1, len(rows))
+	require.Len(t, rows, 1)
 	require.Equal(t, false, createTime == rows[0][4])
 
 	tk.MustExec("create session binding for select * from t using select * from t use index(idx);")
 	rows = tk.MustQuery("show session bindings").Rows()
-	require.Equal(t, 1, len(rows))
+	require.Len(t, rows, 1)
 	createTime = rows[0][4]
 	time.Sleep(1000000)
 	tk.MustExec("create session binding for select * from t using select * from t use index(idx);")
 	rows = tk.MustQuery("show session bindings").Rows()
-	require.Equal(t, 1, len(rows))
+	require.Len(t, rows, 1)
 	require.Equal(t, false, createTime == rows[0][4])
 }
 
 func TestDefaultDB(t *testing.T) {
-	store, _, clean := testkit.CreateMockStoreAndDomain(t)
+	store, clean := testkit.CreateMockStore(t)
 	defer clean()
 
 	tk := testkit.NewTestKit(t, store)
@@ -365,7 +313,7 @@ func TestDefaultDB(t *testing.T) {
 }
 
 func TestIssue19836(t *testing.T) {
-	store, _, clean := testkit.CreateMockStoreAndDomain(t)
+	store, clean := testkit.CreateMockStore(t)
 	defer clean()
 
 	tk := testkit.NewTestKit(t, store)
@@ -391,7 +339,7 @@ func TestIssue19836(t *testing.T) {
 }
 
 func TestTemporaryTable(t *testing.T) {
-	store, _, clean := testkit.CreateMockStoreAndDomain(t)
+	store, clean := testkit.CreateMockStore(t)
 	defer clean()
 
 	tk := testkit.NewTestKit(t, store)
@@ -408,7 +356,7 @@ func TestTemporaryTable(t *testing.T) {
 }
 
 func TestLocalTemporaryTable(t *testing.T) {
-	store, _, clean := testkit.CreateMockStoreAndDomain(t)
+	store, clean := testkit.CreateMockStore(t)
 	defer clean()
 
 	tk := testkit.NewTestKit(t, store)
@@ -424,7 +372,7 @@ func TestLocalTemporaryTable(t *testing.T) {
 }
 
 func TestDropSingleBindings(t *testing.T) {
-	store, _, clean := testkit.CreateMockStoreAndDomain(t)
+	store, clean := testkit.CreateMockStore(t)
 	defer clean()
 
 	tk := testkit.NewTestKit(t, store)
@@ -438,16 +386,16 @@ func TestDropSingleBindings(t *testing.T) {
 	rows := tk.MustQuery("show bindings").Rows()
 	// The size of bindings is equal to one. Because for one normalized sql,
 	// the `create binding` clears all the origin bindings.
-	require.Equal(t, 1, len(rows))
+	require.Len(t, rows, 1)
 	require.Equal(t, "SELECT * FROM `test`.`t` USE INDEX (`idx_b`)", rows[0][1])
 	tk.MustExec("drop binding for select * from t using select * from t use index(idx_a)")
 	rows = tk.MustQuery("show bindings").Rows()
-	require.Equal(t, 1, len(rows))
+	require.Len(t, rows, 1)
 	require.Equal(t, "SELECT * FROM `test`.`t` USE INDEX (`idx_b`)", rows[0][1])
 	tk.MustExec("drop table t")
 	tk.MustExec("drop binding for select * from t using select * from t use index(idx_b)")
 	rows = tk.MustQuery("show bindings").Rows()
-	require.Equal(t, 0, len(rows))
+	require.Len(t, rows, 0)
 
 	tk.MustExec("create table t(a int, b int, c int, index idx_a(a), index idx_b(b))")
 	// Test drop global bindings.
@@ -456,20 +404,20 @@ func TestDropSingleBindings(t *testing.T) {
 	rows = tk.MustQuery("show global bindings").Rows()
 	// The size of bindings is equal to one. Because for one normalized sql,
 	// the `create binding` clears all the origin bindings.
-	require.Equal(t, 1, len(rows))
+	require.Len(t, rows, 1)
 	require.Equal(t, "SELECT * FROM `test`.`t` USE INDEX (`idx_b`)", rows[0][1])
 	tk.MustExec("drop global binding for select * from t using select * from t use index(idx_a)")
 	rows = tk.MustQuery("show global bindings").Rows()
-	require.Equal(t, 1, len(rows))
+	require.Len(t, rows, 1)
 	require.Equal(t, "SELECT * FROM `test`.`t` USE INDEX (`idx_b`)", rows[0][1])
 	tk.MustExec("drop table t")
 	tk.MustExec("drop global binding for select * from t using select * from t use index(idx_b)")
 	rows = tk.MustQuery("show global bindings").Rows()
-	require.Equal(t, 0, len(rows))
+	require.Len(t, rows, 0)
 }
 
 func TestPreparedStmt(t *testing.T) {
-	store, _, clean := testkit.CreateMockStoreAndDomain(t)
+	store, clean := testkit.CreateMockStore(t)
 	defer clean()
 
 	tk := testkit.NewTestKit(t, store)
@@ -485,16 +433,16 @@ func TestPreparedStmt(t *testing.T) {
 	tk.MustExec("create table t(a int, b int, index idx(a))")
 	tk.MustExec(`prepare stmt1 from 'select * from t'`)
 	tk.MustExec("execute stmt1")
-	require.Equal(t, 0, len(tk.Session().GetSessionVars().StmtCtx.IndexNames))
+	require.Len(t, tk.Session().GetSessionVars().StmtCtx.IndexNames, 0)
 
 	tk.MustExec("create binding for select * from t using select * from t use index(idx)")
 	tk.MustExec("execute stmt1")
-	require.Equal(t, 1, len(tk.Session().GetSessionVars().StmtCtx.IndexNames))
+	require.Len(t, tk.Session().GetSessionVars().StmtCtx.IndexNames, 1)
 	require.Equal(t, "t:idx", tk.Session().GetSessionVars().StmtCtx.IndexNames[0])
 
 	tk.MustExec("drop binding for select * from t")
 	tk.MustExec("execute stmt1")
-	require.Equal(t, 0, len(tk.Session().GetSessionVars().StmtCtx.IndexNames))
+	require.Len(t, tk.Session().GetSessionVars().StmtCtx.IndexNames, 0)
 
 	tk.MustExec("drop table t")
 	tk.MustExec("create table t(a int, b int, c int, index idx_b(b), index idx_c(c))")
@@ -502,39 +450,39 @@ func TestPreparedStmt(t *testing.T) {
 
 	tk.MustExec("prepare stmt from 'delete from t where b = ? and c > ?'")
 	tk.MustExec("execute stmt using @p,@p")
-	require.Equal(t, 1, len(tk.Session().GetSessionVars().StmtCtx.IndexNames))
+	require.Len(t, tk.Session().GetSessionVars().StmtCtx.IndexNames, 1)
 	require.Equal(t, "t:idx_b", tk.Session().GetSessionVars().StmtCtx.IndexNames[0])
 	tk.MustExec("create binding for delete from t where b = 2 and c > 2 using delete /*+ use_index(t,idx_c) */ from t where b = 2 and c > 2")
 	tk.MustExec("execute stmt using @p,@p")
-	require.Equal(t, 1, len(tk.Session().GetSessionVars().StmtCtx.IndexNames))
+	require.Len(t, tk.Session().GetSessionVars().StmtCtx.IndexNames, 1)
 	require.Equal(t, "t:idx_c", tk.Session().GetSessionVars().StmtCtx.IndexNames[0])
 
 	tk.MustExec("prepare stmt from 'update t set a = 1 where b = ? and c > ?'")
 	tk.MustExec("execute stmt using @p,@p")
-	require.Equal(t, 1, len(tk.Session().GetSessionVars().StmtCtx.IndexNames))
+	require.Len(t, tk.Session().GetSessionVars().StmtCtx.IndexNames, 1)
 	require.Equal(t, "t:idx_b", tk.Session().GetSessionVars().StmtCtx.IndexNames[0])
 	tk.MustExec("create binding for update t set a = 2 where b = 2 and c > 2 using update /*+ use_index(t,idx_c) */ t set a = 2 where b = 2 and c > 2")
 	tk.MustExec("execute stmt using @p,@p")
-	require.Equal(t, 1, len(tk.Session().GetSessionVars().StmtCtx.IndexNames))
+	require.Len(t, tk.Session().GetSessionVars().StmtCtx.IndexNames, 1)
 	require.Equal(t, "t:idx_c", tk.Session().GetSessionVars().StmtCtx.IndexNames[0])
 
 	tk.MustExec("drop table if exists t1")
 	tk.MustExec("create table t1 like t")
 	tk.MustExec("prepare stmt from 'insert into t1 select * from t where t.b = ? and t.c > ?'")
 	tk.MustExec("execute stmt using @p,@p")
-	require.Equal(t, 1, len(tk.Session().GetSessionVars().StmtCtx.IndexNames))
+	require.Len(t, tk.Session().GetSessionVars().StmtCtx.IndexNames, 1)
 	require.Equal(t, "t:idx_b", tk.Session().GetSessionVars().StmtCtx.IndexNames[0])
 	tk.MustExec("create binding for insert into t1 select * from t where t.b = 2 and t.c > 2 using insert into t1 select /*+ use_index(t,idx_c) */ * from t where t.b = 2 and t.c > 2")
 	tk.MustExec("execute stmt using @p,@p")
-	require.Equal(t, 1, len(tk.Session().GetSessionVars().StmtCtx.IndexNames))
+	require.Len(t, tk.Session().GetSessionVars().StmtCtx.IndexNames, 1)
 	require.Equal(t, "t:idx_c", tk.Session().GetSessionVars().StmtCtx.IndexNames[0])
 
 	tk.MustExec("prepare stmt from 'replace into t1 select * from t where t.b = ? and t.c > ?'")
 	tk.MustExec("execute stmt using @p,@p")
-	require.Equal(t, 1, len(tk.Session().GetSessionVars().StmtCtx.IndexNames))
+	require.Len(t, tk.Session().GetSessionVars().StmtCtx.IndexNames, 1)
 	require.Equal(t, "t:idx_b", tk.Session().GetSessionVars().StmtCtx.IndexNames[0])
 	tk.MustExec("create binding for replace into t1 select * from t where t.b = 2 and t.c > 2 using replace into t1 select /*+ use_index(t,idx_c) */ * from t where t.b = 2 and t.c > 2")
 	tk.MustExec("execute stmt using @p,@p")
-	require.Equal(t, 1, len(tk.Session().GetSessionVars().StmtCtx.IndexNames))
+	require.Len(t, tk.Session().GetSessionVars().StmtCtx.IndexNames, 1)
 	require.Equal(t, "t:idx_c", tk.Session().GetSessionVars().StmtCtx.IndexNames[0])
 }
